move extrinsic item content component to own file
